pushing deletion confirmation

+++ vue/src/components/ItineraryList.vue
@@ -11,7 +11,7 @@
         Itinerary ID: {{itinerary.itineraryId}} <br>
         <p>***MAP GOES HERE***</p>
         <router-link :to="{ name: 'itinerary-details', params: {id: itinerary.itineraryId}}"> Edit Itinerary </router-link><br><br>
-        <button v-on:click.prevent='deleteItinerary(itinerary.itineraryId)' onClick="window.location.reload();"> Delete </button>
+        <button v-on:click.prevent='deleteItinerary(itinerary.itineraryId)'> Delete </button>
     </div>
 
     <h2>Itineraries Shared With You</h2>
@@ -69,7 +69,15 @@ export default {
             }
         },
         deleteItinerary(id) {
-            itineraryService.deleteItinerary(id);
+            if (
+                confirm("Are you sure you want to delete this itinerary? This action cannot be undone.")
+            ) {
+                itineraryService.deleteItinerary(id)
+                
+            }
+            window.location.reload().then(
+                    alert("Itinerary successfully deleted."));
+            
         },
         getItineraryLandmarks(id){
             itineraryService.getItineraryLandmarks(id);
